requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
argcomplete==1.10.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.8.2
blinker==1.6.2
certifi==2023.5.7
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
compressed-rtf==1.0.6
debugpy==1.6.7
decorator==5.1.1
docx2txt==0.8
ebcdic==1.1.1
executing==1.2.0
extract-msg==0.28.7
Flask==2.3.2
frozenlist==1.3.3
idna==3.4
IMAPClient==2.1.0
importlib-metadata==6.7.0
ipykernel==6.23.3
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.3.0
jupyter_core==5.3.1
lxml==4.9.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.24.2
olefile==0.46
openai==0.27.1
packaging==23.1
pandas==1.5.3
parso==0.8.3
pdfminer.six==20191110
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.8.0
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
pycryptodome==3.18.0
Pygments==2.15.1
python-dateutil==2.8.2
python-pptx==0.6.21
pytz==2023.3
pywin32==306
pyzmq==25.1.0
redis==4.5.1
regex==2023.6.3
requests==2.28.2
six==1.12.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SpeechRecognition==3.8.1
stack-data==0.6.2
textract==1.6.5
tiktoken==0.4.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.7.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.16
wcwidth==0.2.6
Werkzeug==2.3.6
xlrd==1.2.0
XlsxWriter==3.1.2
yarl==1.9.2
zipp==3.15.0
